+++ src/components/offer/offer-review-form.tsx
@@ -1,0 +1,144 @@
+import { useState } from 'react';
+
+function OfferReviewForm(): JSX.Element {
+  const [rating, setRating] = useState(0);
+  const [review, setReview] = useState('');
+
+  function handleRatingChange(e) {
+    setRating(e.target.value);
+  }
+
+  function handleReviewChange(e) {
+    setReview(e.target.value);
+  }
+
+  return (
+    <form className="reviews__form form" action="#" method="post">
+      <label className="reviews__label form__label" htmlFor="review">
+            Your review
+      </label>
+      <div className="reviews__rating-form form__rating">
+        <input
+          value={rating}
+          onChange={handleRatingChange}
+          className="form__rating-input visually-hidden"
+          name="rating"
+          defaultValue={5}
+          id="5-stars"
+          type="radio"
+        />
+        <label
+          htmlFor="5-stars"
+          className="reviews__rating-label form__rating-label"
+          title="perfect"
+        >
+          <svg className="form__star-image" width={37} height={33}>
+            <use xlinkHref="#icon-star" />
+          </svg>
+        </label>
+        <input
+          value={rating}
+          onChange={handleRatingChange}
+          className="form__rating-input visually-hidden"
+          name="rating"
+          defaultValue={4}
+          id="4-stars"
+          type="radio"
+        />
+        <label
+          htmlFor="4-stars"
+          className="reviews__rating-label form__rating-label"
+          title="good"
+        >
+          <svg className="form__star-image" width={37} height={33}>
+            <use xlinkHref="#icon-star" />
+          </svg>
+        </label>
+        <input
+          value={rating}
+          onChange={handleRatingChange}
+          className="form__rating-input visually-hidden"
+          name="rating"
+          defaultValue={3}
+          id="3-stars"
+          type="radio"
+        />
+        <label
+          htmlFor="3-stars"
+          className="reviews__rating-label form__rating-label"
+          title="not bad"
+        >
+          <svg className="form__star-image" width={37} height={33}>
+            <use xlinkHref="#icon-star" />
+          </svg>
+        </label>
+        <input
+          value={rating}
+          onChange={handleRatingChange}
+          className="form__rating-input visually-hidden"
+          name="rating"
+          defaultValue={2}
+          id="2-stars"
+          type="radio"
+        />
+        <label
+          htmlFor="2-stars"
+          className="reviews__rating-label form__rating-label"
+          title="badly"
+        >
+          <svg className="form__star-image" width={37} height={33}>
+            <use xlinkHref="#icon-star" />
+          </svg>
+        </label>
+        <input
+          value={rating}
+          onChange={handleRatingChange}
+          className="form__rating-input visually-hidden"
+          name="rating"
+          defaultValue={1}
+          id="1-star"
+          type="radio"
+        />
+        <label
+          htmlFor="1-star"
+          className="reviews__rating-label form__rating-label"
+          title="terribly"
+        >
+          <svg className="form__star-image" width={37} height={33}>
+            <use xlinkHref="#icon-star" />
+          </svg>
+        </label>
+      </div>
+      <textarea
+        value={review}
+        onChange={handleReviewChange}
+        className="reviews__textarea form__textarea"
+        id="review"
+        name="review"
+        placeholder="Tell how was your stay, what you like and what can be improved"
+        defaultValue={''}
+      />
+      <div className="reviews__button-wrapper">
+        <p className="reviews__help">
+              To submit review please make sure to set{' '}
+          <span className="reviews__star">rating</span> and describe
+              your stay with at least{' '}
+          <b className="reviews__text-amount">50 characters</b>.
+        </p>
+        <button
+          onClick={() =>{
+            setRating(1);
+            setReview(review);
+          }}
+          className="reviews__submit form__submit button"
+          type="submit"
+          disabled={undefined}
+        >
+              Submit
+        </button>
+      </div>
+    </form>
+  );
+}
+
+export default OfferReviewForm;
