+++ src/index.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './app/app';
 import { Setting } from './const';
-import { offersList, сurrentOffer } from './mock/offers';
+import { offersList } from './mock/offers';
+import { сurrentOffer } from './mock/сurrentOffer';
 import { comments } from './mock/comments';
 
 const root = ReactDOM.createRoot(
