+++ src/mock/offers.ts
@@ -1,4 +1,4 @@
-import { OffersList, CurrentOffer } from '../types/offer';
+import { OffersList } from '../types/offer';
 
 export const offersList: OffersList[] = [
   {
@@ -92,51 +92,4 @@ export const offersList: OffersList[] = [
     'isPremium': true,
     'rating': 5
   },
-];
-
-export const сurrentOffer: CurrentOffer[] = [
-  {
-    'id': '6af6f711-c28d-4121-82cd-27f00e0b462a',
-    'title': 'Beautiful & luxurious studio at great location',
-    'description': 'Relax, rejuvenate and unplug in this ultimate rustic getaway experience in the country. In our beautiful screened Pondhouse, you can gaze at the stars and listen to the sounds of nature from your cozy warm bed.',
-    'type': 'apartment',
-    'price': 120,
-    'images': [
-      'https://15.design.htmlacademy.pro/static/hotel/18.jpg',
-      'https://15.design.htmlacademy.pro/static/hotel/6.jpg',
-      'https://15.design.htmlacademy.pro/static/hotel/10.jpg',
-      'https://15.design.htmlacademy.pro/static/hotel/14.jpg',
-      'https://15.design.htmlacademy.pro/static/hotel/4.jpg',
-      'https://15.design.htmlacademy.pro/static/hotel/17.jpg'
-    ],
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      },
-    },
-    'location': {
-      'latitude': 52.35514938496378,
-      'longitude': 4.673877537499948,
-      'zoom': 8
-    },
-    'goods': [
-      'Heating',
-      'Laptop friendly workspace',
-      'Cable TV',
-      'Fridge'
-    ],
-    'host': {
-      'isPro': true,
-      'name': 'Angelina',
-      'avatarUrl': 'https://15.design.htmlacademy.pro/static/host/avatar-angelina.jpg'
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4,
-    'bedrooms': 3,
-    'maxAdults': 3
-  }
 ];
