+++ src/mock/сurrentOffer.ts
@@ -1,0 +1,188 @@
+import { CurrentOffer } from '../types/offer';
+
+export const сurrentOffer: CurrentOffer[] = [
+  {
+    'id': '6af6f711-c28d-4121-82cd-e0b462a27f00',
+    'title': 'Beautiful & luxurious studio at great location',
+    'description': 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget.',
+    'type': 'apartment',
+    'price': 120,
+    'images': [
+      'https://loremflickr.com/248/152?random=10',
+      'https://loremflickr.com/248/152?random=11',
+      'https://loremflickr.com/248/152?random=12',
+      'https://loremflickr.com/248/152?random=13',
+      'https://loremflickr.com/248/152?random=14',
+      'https://loremflickr.com/248/152?random=15',
+    ],
+    'city': {
+      'name': 'Amsterdam',
+      'location': {
+        'latitude': 52.35514938496378,
+        'longitude': 4.673877537499948,
+        'zoom': 8
+      }},
+    'location': {
+      'latitude': 52.35514938496378,
+      'longitude': 4.673877537499948,
+      'zoom': 8
+    },
+    'goods': [
+      'Heating',
+      'Laptop friendly workspace',
+      'Cable TV',
+      'Fridge'
+    ],
+    'host': {
+      'isPro': true,
+      'name': 'Angelina',
+      'avatarUrl': 'https://loremflickr.com/248/152?random=16'
+    },
+    'isFavorite': true,
+    'isPremium': false,
+    'rating': 4,
+    'bedrooms': 3,
+    'maxAdults': 3
+  },
+  {
+    'id': '2f47ca94-4d24-4425-9ae6-43d7333df736',
+    'title': 'House in countryside',
+    'description': 'Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis',
+    'type': 'apartment',
+    'price': 242,
+    'images': [
+      'https://loremflickr.com/248/152?random=20',
+      'https://loremflickr.com/248/152?random=21',
+      'https://loremflickr.com/248/152?random=22',
+      'https://loremflickr.com/248/152?random=23',
+      'https://loremflickr.com/248/152?random=24',
+      'https://loremflickr.com/248/152?random=25',
+    ],
+    'city': {
+      'name': 'Brussels',
+      'location': {
+        'latitude': 48.85661,
+        'longitude': 2.351499,
+        'zoom': 13
+      }
+    },
+    'location': {
+      'latitude': 48.858610000000006,
+      'longitude': 2.330499,
+      'zoom': 16
+    },
+    'goods': [
+      'Wi-Fi',
+      'Washing machine',
+      'Towels',
+      'Heating',
+      'Coffee machine',
+      'Baby seat',
+      'Kitchen',
+      'Dishwasher',
+      'Cabel TV',
+      'Fridge',
+    ],
+    'host': {
+      'isPro': false,
+      'name': 'Jack',
+      'avatarUrl': 'https://loremflickr.com/248/152?random=26'
+    },
+    'isFavorite': true,
+    'isPremium': true,
+    'rating': 2.7,
+    'bedrooms': 1,
+    'maxAdults': 2
+  },
+  {
+    'id': '258e916e-b073-49d5-8806-e4e2fb3617a3',
+    'title': 'Tile House',
+    'description': 'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.',
+    'type': 'apartment',
+    'price': 101,
+    'images': [
+      'https://loremflickr.com/248/152?random=30',
+      'https://loremflickr.com/248/152?random=31',
+      'https://loremflickr.com/248/152?random=32',
+      'https://loremflickr.com/248/152?random=33',
+      'https://loremflickr.com/248/152?random=34',
+      'https://loremflickr.com/248/152?random=35',
+    ],
+    'city': {
+      'name': 'Paris',
+      'location': {
+        'latitude': 48.85661,
+        'longitude': 2.351499,
+        'zoom': 13
+      }
+    },
+    'location': {
+      'latitude': 48.834610000000005,
+      'longitude': 2.335499,
+      'zoom': 16
+    },
+    'goods': [
+      'Towels',
+      'Coffee machine',
+      'Baby seat',
+      'Dishwasher',
+      'Fridge',
+    ],
+    'host': {
+      'isPro': true,
+      'name': 'Niko',
+      'avatarUrl': 'https://loremflickr.com/248/152?random=36'
+    },
+    'isFavorite': false,
+    'isPremium': false,
+    'rating': 2,
+    'bedrooms': 2,
+    'maxAdults': 5
+  },
+  {
+    'id': 'f431d956-6b7f-4aa4-839e-d57743638b28',
+    'title': 'Tile House',
+    'description': 'Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus.',
+    'type': 'room',
+    'price': 216,
+    'images': [
+      'https://loremflickr.com/248/152?random=40',
+      'https://loremflickr.com/248/152?random=41',
+      'https://loremflickr.com/248/152?random=42',
+      'https://loremflickr.com/248/152?random=43',
+      'https://loremflickr.com/248/152?random=44',
+      'https://loremflickr.com/248/152?random=45',
+    ],
+    'city': {
+      'name': 'Amsterdam',
+      'location': {
+        'latitude': 48.85661,
+        'longitude': 2.351499,
+        'zoom': 13
+      }
+    },
+    'location': {
+      'latitude': 48.85761,
+      'longitude': 2.358499,
+      'zoom': 16
+    },
+    'goods': [
+      'Wi-Fi',
+      'Heating',
+      'Coffee machine',
+      'Kitchen',
+      'Dishwasher',
+      'Fridge',
+    ],
+    'host': {
+      'isPro': false,
+      'name': 'Jane',
+      'avatarUrl': 'https://loremflickr.com/248/152?random=46'
+    },
+    'isFavorite': true,
+    'isPremium': true,
+    'rating': 5,
+    'bedrooms': 5,
+    'maxAdults': 15
+  },
+];
